+++ CHANGELOG.rst
@@ -14,6 +14,8 @@ Change Log
 Unreleased
 ~~~~~~~~~~
 
+[2.5.5] - 2021-01-05
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 * Cover `Start System Check` button on the proctoring instruction page with the
   conditions software download link is provided by the proctoring provider,
   since some providers do not has that step in the onboarding process.
@@ -21,6 +23,7 @@ Unreleased
 * Added `time_remaining_seconds` field to the exam attempt model in order to
   allow the remaining time on an exam attempt to be saved after it enters an
   error state.
+* Fix bug allowing learners access to onboarding setup after exam due date.
 
 [2.5.4] - 2020-12-17
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~